remove uneeded mem usage

+++ src/engine/battle/Battle.cpp
@@ -147,7 +147,7 @@ void BattleEngine::opponentActionFirst() {
 }
 
 void BattleEngine::changeCurMon(uint8_t id) {
-    int index = 0;
+    uint8_t index = 0;
     for (uint8_t i = 0; i < 3; i++) {
         if (this->playerParty[i]->id == id) {
             this->playerCur = this->playerParty[i];
@@ -172,7 +172,7 @@ void BattleEngine::changeCurMon(uint8_t id) {
 }
 
 bool BattleEngine::tryCapture() {
-    int roll = random(1, 10);
+    uint8_t roll = random(1, 10);
     return roll < 5;
 }
